Pass requirment to Source#infer_from_opts
So that we can pass the requirement to the Source objects themselves.

With the current code, I can't see how it was planned to work as the version
to be installed was calculated in the PuppetModule instance but never passed
to the Source object which actually handles the fetching, installing, etc.